split non-gemm specific generator code to gen_tensor_op.py
commit 250f915652e72e0012e9aa6ce0b6ef337d3da845
Author: Masahiro Masuda <masahi129@gmail.com>
Date:   Sun Nov 14 06:44:52 2021 +0900

    remove conv2d stuff

commit 1a6b27c438472f13acd4a0f466d78f293415e076
Author: Masahiro Masuda <masahi129@gmail.com>
Date:   Sun Nov 14 06:41:31 2021 +0900

    remove unused import

commit f7c3b5a191b8c73e8b178c32f6d3182fb0f697d6
Author: Masahiro Masuda <masahi129@gmail.com>
Date:   Sun Nov 14 06:37:07 2021 +0900

    add profiler boilarplate for conv2d

commit ca1ae274fb8f96a1dcde688deaf15339fe5604fb
Author: Masahiro Masuda <masahi129@gmail.com>
Date:   Sun Nov 14 06:22:06 2021 +0900

    introduce gen_tensor_op.py

commit 37bb918e0873f04457c29479eb21a530b7052217
Author: Masahiro Masuda <masahi129@gmail.com>
Date:   Sun Nov 14 05:45:41 2021 +0900

    more conv2d code

commit 5c00398892c99cb2a03be51f75878992663432dd
Author: Masahiro Masuda <masahi129@gmail.com>
Date:   Sun Nov 14 05:13:30 2021 +0900

    Begin conv2d support
